requirements.txt
fastapi
uvicorn[standard]
s3fs
google-cloud-bigquery
google-auth
db-dtypes
gcloud